+++ app/src/main/java/com/example/valorgasto/ui/purchase_form/ProductForm.kt
@@ -167,7 +167,7 @@ fun ProductForm(
                                     name = productName,
                                     quantity = productQuantity.toInt(),
                                     unitOfMeasurement = selectedUnitOfMeasurement,
-                                    price = productPrice.toLong(),
+                                    price = productPrice.toFloat(),
                                     category = selectedCategory
                                 )
                                 onConfirm(product)
